cmd/combine: move combine package
